src/requirements.txt
pyTelegramBotAPI==4.14.0
numpy
google-cloud-storage
google-cloud-datastore
google-cloud-logging